+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.0.1] - 2025-02-06
+## [0.0.2] - 2025-06-12
+
+### Changes
+
+- Separate class **PersistenceLayerSetup** (_setup.py_) for setup persistence layer that is dependable on the engine
+- FlatFile added as a Persistence Engine (**FlatFilePersistenceEngine** class)
+- **add_content** method added to Persistence Engine
+- **CLI** added for the app (for now based on argparse)
+
+## [0.0.1] - 2025-06-08
 
 ### Changes